Add final modifications

+++ Unit/src/Header Files/sensors.h
@@ -1,10 +1,12 @@
 #ifndef SENSORS_H_
 #define SENSORS_H_
 
+#include <avr/io.h>
+
 #define INVALID_READING_VALUE -128
 
 float getTemperature();
 float getLightIntensity();
-float getDistance(void);
+uint8_t getDistance(void);
 
 #endif /* SENSORS_H_ */
